Move phone collect flow into common flows directory
